% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{.get_admin_levels_data}
\alias{.get_admin_levels_data}
\title{Obtener datos de todos los niveles administrativos para detección}
\usage{
.get_admin_levels_data()
}
\value{
Una lista con los datos de cada nivel administrativo.
}
\description{
Función auxiliar que obtiene los datos de referencia de todos los niveles
administrativos disponibles en geodomR.
}
\keyword{internal}
