% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bparajes.R
\name{gd_bparajes}
\alias{gd_bparajes}
\title{Obtener Límites de los Barrios y Parajes de la República Dominicana}
\usage{
gd_bparajes(id = "RD_BPARAJES", sf = TRUE)
}
\arguments{
\item{id}{Nombre del archivo de datos en el servidor remoto.}

\item{sf}{Lógico. Si es \code{FALSE}, devuelve un data.frame regular sin la
columna de geometría. Por defecto es \code{TRUE}.}
}
\value{
Un objeto de la clase \code{sf} o un \code{data.frame}.
}
\description{
Descarga (si es necesario) y carga los límites de los barrios y parajes de la
República Dominicana, según la division correspondiente al id suministrado.
}
\details{
Las divisiones disponibles son:
\itemize{
\item \code{"RD_BPARAJES"} (por defecto): Barrios y Parajes de la República Dominicana.
}
}
\examples{
\dontrun{
  # Cargar el objeto sf completo
  bparajes_sf <- gd_bparajes()

  # Cargar solo la tabla de atributos (sin geometría)
  bparajes_df <- gd_bparajes(sf = FALSE)
}
}
