% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{gd_detect_level}
\alias{gd_detect_level}
\title{Detectar el nivel administrativo de datos geográficos}
\usage{
gd_detect_level(data, .level = NULL, .name = NULL, .key = NULL)
}
\arguments{
\item{data}{Un data frame que contiene los datos a analizar.}

\item{.level}{Cadena de caracteres opcional especificando el nivel administrativo.
Si se proporciona, la función solo considerará este nivel. Opciones válidas son:
"sections", "dm", "municipalities", "provinces", "regions", "bparajes".}

\item{.name}{Cadena de caracteres opcional especificando el nombre de la variable
en \code{data} que contiene los nombres de las unidades administrativas. Si se
proporciona, la función solo considerará esta variable.}

\item{.key}{Cadena de caracteres opcional especificando el nombre de la variable
clave en los datos de límites administrativos. Si se proporciona, la función
solo considerará esta clave.}
}
\value{
Una lista que contiene los siguientes elementos:
\itemize{
\item \code{level}: El nivel administrativo inferido.
\item \code{name}: El nombre de la variable en \code{data} que contiene los nombres
de las unidades administrativas.
\item \code{key}: El nombre de la variable clave en los datos de límites administrativos.
\item \code{match_count}: El número de valores únicos que coincidieron.
\item \code{total_count}: El número total de valores únicos en la columna.
}
}
\description{
Esta función intenta determinar automáticamente el nivel administrativo
apropiado, el nombre de la variable de enlace, y la variable clave para
datos relacionados con la República Dominicana. Examina los datos
proporcionados y los compara con los límites administrativos disponibles
para encontrar la mejor coincidencia.
}
\examples{
\dontrun{
  # Detectar información para datos de municipios
  mis_datos <- data.frame(municipio = c("Santo Domingo Este", "Santiago"), valor = 1:2)
  info <- gd_detect_level(mis_datos)

  # Especificar el nivel administrativo
  info <- gd_detect_level(mis_datos, .level = "municipalities")
}
}
