% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{gd_dm}
\alias{gd_dm}
\title{Obtener Límites de los Distritos Municipales de la República Dominicana}
\usage{
gd_dm(id = "RD_DM", sf = TRUE)
}
\arguments{
\item{id}{Nombre del archivo de datos en el servidor remoto.}

\item{sf}{Lógico. Si es \code{FALSE}, devuelve un data.frame regular sin la
columna de geometría. Por defecto es \code{TRUE}.}
}
\value{
Un objeto de la clase \code{sf} o un \code{data.frame}.
}
\description{
Descarga (si es necesario) y carga los límites de los distritos municipales
de la República Dominicana, según la division correspondiente al id suministrado.
}
\details{
Las divisiones disponibles son:
\itemize{
\item \code{"RD_DM"} (por defecto): Distritos Municipales de la República Dominicana.
}
}
\examples{
\dontrun{
  # Cargar el objeto sf completo
  distritos_sf <- gd_dm()

  # Cargar solo la tabla de atributos (sin geometría)
  distritos_df <- gd_dm(sf = FALSE)
}
}
