% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{gd_get_dataset}
\alias{gd_get_dataset}
\title{Descargar un dataset comunitario de GeoDOM}
\usage{
gd_get_dataset(id, force_download = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{id}{Nombre del dataset, sin la extensión JSON.}

\item{force_download}{Si es TRUE, descarga nuevamente el dataset.}

\item{verbose}{Si es TRUE, muestra mensajes de progreso.}

\item{...}{Reservado para compatibilidad.}
}
\value{
Una lista con los metadatos y las filas en \code{data}.
}
\description{
Descargar un dataset comunitario de GeoDOM
}
