% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroregions.R
\name{gd_macroregions}
\alias{gd_macroregions}
\title{Obtener Límites de las Macro-Regiones}
\usage{
gd_macroregions(id = "RD_MREG", sf = TRUE)
}
\arguments{
\item{id}{Nombre del archivo de datos en el servidor remoto.}

\item{sf}{Lógico. Si es \code{FALSE}, devuelve un data.frame regular sin la
columna de geometría. Por defecto es \code{TRUE}.}
}
\value{
Un objeto de la clase \code{sf} o un \code{data.frame}.
}
\description{
Descarga (si es necesario) y carga los límites de las tres macro-regiones
de planificación como un objeto \code{sf}.
}
\examples{
\dontrun{
  # Cargar el objeto sf completo
  macro_regiones_sf <- gd_macroregions()

  # Cargar solo la tabla de atributos (sin geometría)
  macro_regiones_df <- gd_macroregions(sf = FALSE)
}
}
