% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provinces.R
\name{gd_provinces}
\alias{gd_provinces}
\title{Obtener Límites de las Provincias de la República Dominicana}
\usage{
gd_provinces(id = "RD_PROV", sf = TRUE, .reg = NULL)
}
\arguments{
\item{id}{Nombre del archivo de datos en el servidor remoto. Por defecto es \code{"RD_PROV"}.}

\item{sf}{Lógico. Si es \code{FALSE}, devuelve un \code{data.frame} regular sin la
columna de geometría. Por defecto es \code{TRUE}.}

\item{.reg}{Caracter. Si es \code{"rup"}, agrega la columna de región administrativa
según la Ley 345-22. Por defecto es \code{NULL}.}
}
\value{
Un objeto de la clase \code{sf} o un \code{data.frame}, dependiendo del valor de \code{sf}.
Si \code{.reg = "rup"}, incluye la columna de región administrativa.
}
\description{
Descarga (si es necesario) y carga los límites de las provincias de la
República Dominicana. Permite obtener los datos como objeto \code{sf} o como
\code{data.frame} sin geometría. Además, puede agregar la columna de región
administrativa según la Ley 345-22 si se especifica el argumento \code{.reg = "rup"}.
}
\examples{
\dontrun{
  # Cargar el objeto sf completo
  provincias_sf <- gd_provinces()

  # Cargar solo la tabla de atributos (sin geometría)
  provincias_df <- gd_provinces(sf = FALSE)

  # Cargar provincias con columna de región administrativa
  provincias_reg <- gd_provinces(.reg = "rup")
}
}
