% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{gd_regions}
\alias{gd_regions}
\title{Obtener Límites de las Regiones de Planificación}
\usage{
gd_regions(id = "RD_RUP", sf = TRUE, verbose = FALSE)
}
\arguments{
\item{id}{Nombre del archivo de datos en el servidor remoto.}

\item{sf}{Lógico. Si es \code{FALSE}, devuelve un data.frame regular sin la
columna de geometría. Por defecto es \code{TRUE}.}

\item{verbose}{Lógico. Si es \code{TRUE}, muestra mensajes informativos durante la
descarga y carga de datos. Por defecto es \code{FALSE}.}
}
\value{
Un objeto de la clase \code{sf} o un \code{data.frame}.
}
\description{
Descarga (si es necesario) y carga los límites de las regiones de
planificación según la regionalización correspondiente al id suministrado.
}
\details{
Las regionalizaciones disponibles son:
\itemize{
\item \code{"RD_RUP"} (por defecto): Regiones Únicas de Planificación según Ley 345-22.
\item \code{"RD_REG71004"}: Regiones de planificación según el Decreto 710-04.
}
}
\examples{
\dontrun{
# Cargar el objeto sf completo
regiones_sf <- gd_regions()

# Cargar solo la tabla de atributos (sin geometría)
regiones_df <- gd_regions(sf = FALSE)
}
}
