% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{gd_sections}
\alias{gd_sections}
\title{Obtener Límites de las Secciones de la República Dominicana}
\usage{
gd_sections(id = "RD_SECCIONES", sf = TRUE)
}
\arguments{
\item{id}{Nombre del archivo de datos en el servidor remoto.}

\item{sf}{Lógico. Si es \code{FALSE}, devuelve un data.frame regular sin la
columna de geometría. Por defecto es \code{TRUE}.}
}
\value{
Un objeto de la clase \code{sf} o un \code{data.frame}.
}
\description{
Descarga (si es necesario) y carga los límites de las secciones de la
República Dominicana, según la division correspondiente al id suministrado.
}
\details{
Las divisiones disponibles son:
\itemize{
\item \code{"RD_SECCIONES"} (por defecto): Secciones de la República Dominicana.
}
}
\examples{
\dontrun{
  # Cargar el objeto sf completo
  secciones_sf <- gd_sections()

  # Cargar solo la tabla de atributos (sin geometría)
  secciones_df <- gd_sections(sf = FALSE)
}
}
