% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\name{gd_clean_municipality_name}
\alias{gd_clean_municipality_name}
\title{Limpia y estandariza los nombres de municipios de la República Dominicana}
\usage{
gd_clean_municipality_name(mun, .tol = 0.25, .on_error = "fail")
}
\arguments{
\item{mun}{Vector de caracteres con los nombres de municipios a limpiar.}

\item{.tol}{Nivel de tolerancia numérica para la similitud de cadenas. Por defecto es 0.25.
Este parámetro controla cuán similares deben ser dos cadenas para considerarse una coincidencia.
Un valor más bajo significa una coincidencia más estricta.}

\item{.on_error}{Cadena de caracteres que especifica el método de manejo de errores. Por defecto es "fail".
Puede ser uno de los siguientes: "fail" para detener la ejecución en caso de error,
"omit" para ignorar los nombres no coincidentes, o "na" para devolver NA en los nombres no coincidentes.}
}
\value{
Un vector de caracteres con los nombres de municipios limpiados.
}
\description{
Esta función limpia y estandariza los nombres de los municipios en la República Dominicana,
con tolerancia para la similitud de cadenas y opciones para el manejo de errores. Utiliza
el sistema oficial de códigos municipales de la República Dominicana.
}
\examples{
\dontrun{
# Uso básico con nombres de municipios
nombres_mun_limpios <- gd_clean_municipality_name(c("santiago", "moca", "bonao"))

# Con variantes de prefijos
gd_clean_municipality_name(c("Municipio de Santiago", "Mun. Moca"))

# Con tolerancia y manejo de errores
gd_clean_municipality_name("santiagooo", .tol = 0.8, .on_error = "na")
}
}
