% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\name{gd_municipalities}
\alias{gd_municipalities}
\title{Obtener Límites de los Municipios de la República Dominicana}
\usage{
gd_municipalities(id = "RD_MUN158", sf = TRUE)
}
\arguments{
\item{id}{Nombre del archivo de datos en el servidor remoto.}

\item{sf}{Lógico. Si es \code{FALSE}, devuelve un data.frame regular sin la
columna de geometría. Por defecto es \code{TRUE}.}
}
\value{
Un objeto de la clase \code{sf} o un \code{data.frame}.
}
\description{
Descarga (si es necesario) y carga los límites de los municipios de la
República Dominicana, según la division correspondiente al id suministrado.
}
\details{
Las divisiones disponibles son:
\itemize{
\item \code{"RD_MUN158"} (por defecto): 158 Municipios de la República Dominicana.
\item \code{"RD_MUN155"}: TODO
}
}
\examples{
\dontrun{
# Cargar el objeto sf completo
municipios_sf <- gd_municipalities()

# Cargar solo la tabla de atributos (sin geometría)
municipios_df <- gd_municipalities(sf = FALSE)
}
}
